Configures Play's Require.Js support to correctly build / minify.
Note: there are still problems with this configuration:

a) Third-party libraries (jquery, underscore, ...) are minified.

I'd rather provide their own minified versions myself.
Play's minified version of jQuery is ~96kb, against ~32kb for the official.

Also, its really slow to minify those...

b) Inlined modules are not optimal

Consider the inlined layout of both versions:
The Play! version in the left, and the r.js version to the right.

    common.js                          custom build.js
    --------------------------------------------------
    common.js                   common.js
                                app/lib.js
                                app/controller/Base.js
                                app/model/Base.js

    app/main1.js
    --------------------------------------------------
    app/lib.js
    app/controller/Base.js
    app/controller/c1.js        app/controller/c1.js
    app/model/Base.js
    app/model/m1.js             app/model/m1.js
    app/main1.js                app/main1.js

    app/main2.js
    --------------------------------------------------
    app/lib.js
    app/controller/Base.js
    app/controller/c2.js        app/controller/c2.js
    app/model/Base.js
    app/model/m2.js             app/model/m2.js
    app/main2.js                app/main2.js